feat: create generic index files
