+++ template/base/src/pages/index.tsx
@@ -1,6 +1,12 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 
+type TechnologyCardProps = {
+  name: string;
+  description: string;
+  documentation: string;
+};
+
 const Home: NextPage = () => {
   return (
     <>
@@ -9,33 +15,40 @@ const Home: NextPage = () => {
         <meta name="description" content="Generated by create-t3-app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <div>
+
+      <main>
         <h1>
           Create <span>T3</span> App
         </h1>
-
-        <div>
-          <h3>This stack uses:</h3>
-          <ul>
-            <li>
-              <a href="https://nextjs.org" target="_blank" rel="noreferrer">
-                Next.js
-              </a>
-            </li>
-            <li>
-              <a
-                href="https://typescriptlang.org"
-                target="_blank"
-                rel="noreferrer"
-              >
-                TypeScript
-              </a>
-            </li>
-          </ul>
-        </div>
-      </div>
+        <p>This stack uses:</p>
+        <ul>
+          <TechnologyCard
+            name="NextJS"
+            description="The React framework for production"
+            documentation="https://nextjs.org/"
+          />
+          <TechnologyCard
+            name="TypeScript"
+            description="Strongly typed programming language that builds on JavaScript, giving you better tooling at any scale"
+            documentation="https://www.typescriptlang.org/"
+          />
+        </ul>
+      </main>
     </>
   );
 };
+
+const TechnologyCard = ({
+  name,
+  description,
+  documentation,
+}: TechnologyCardProps) => (
+  <li>
+    <a href={documentation} target="_blank" rel="noreferrer">
+      {name}
+    </a>
+    <span> - {description}</span>
+  </li>
+);
 
 export default Home;
