+++ template/base/src/styles/globals.css
@@ -8,7 +8,6 @@ body {
 
 a {
   color: inherit;
-  text-decoration: none;
 }
 
 * {
